kafka/EventAggregator/requirements.txt
asyncio==3.4.3
kafka-python==2.0.2
faust-streaming==0.10.22